// HealthCare/src/Interfaces/Employee/employee-response.ts
export interface EmployeeResponse {
  recordId: number;
  employeeId: string;
  employeeName: string;
  emailId: string;
  dateOfJoining: string;
  contactNumber: string;
  mobileNumber: string;
  department: string;
  designation: string;
  qualification: string;
  recordStatus: boolean;
  address: string;
  isPathology: boolean;
  signatureImage: string;
  partnerId: string;
  createdBy: string;
}