// HealthCare/src/Interfaces/Employee/employee-request.ts
export interface EmployeeRequest {
  employeeId: string;
  employeeName: string;
  emailId: string;
  dateOfJoining: Date;
  contactNumber: string;
  mobileNumber: string;
  department: string;
  designation: string;
  qualification: string;
  recordStatus: boolean;
  address: string;
  isPathology: boolean;
  signatureImage: string;
  partnerId: string;
  createdBy: string;
}